% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Bet}
\alias{Bet}
\title{Make predictions with the trained SciBet model.}
\usage{
Bet(expr, result=c("list", "table"))
}
\arguments{
\item{expr}{The expression matrix or dataframe for prediction. Rows should be cells and columns should be genes.}

\item{result}{Return a "list" of predicted labels, or a "table" of probabilities of each tested cell belonging to each label.}
}
\value{
A vector or a dataframe for the classification result.
}
\description{
Make predictions with the trained SciBet model.
}
