% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{conf_score}
\alias{conf_score}
\title{Calculating the confidence score C for false positive control.}
\usage{
conf_score(ref, query, null_expr, gene_num = 500)
}
\arguments{
\item{ref}{The reference dataset. Rows should be cells, columns should be genes and last column should be "label".}

\item{query}{The query dataset. Rows should be cells and columns should be genes.}

\item{gene_num}{The number of common markers of reference set used for false positive control.}
}
\value{
A vector of confidence scores.
}
\description{
Calculating the confidence score C for false positive control.
}
