% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Confusion_heatmap}
\alias{Confusion_heatmap}
\title{Heatmap of classification result.}
\usage{
Confusion_heatmap(ori, prd)
}
\arguments{
\item{ori}{A vector of the original labels for each cell in the test set.}

\item{prd}{A vector of the predicted labels for each cell in the test set..}
}
\value{
A heatmap for the confusion matrix of the classification result.
}
\description{
Heatmap of classification result.
}
