% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Confusion_heatmap_negctrl}
\alias{Confusion_heatmap_negctrl}
\title{Heatmap for the confusion matrix of the classification with the false postive control.}
\usage{
Confusion_heatmap_negctrl(res, cutoff = 0.4)
}
\arguments{
\item{res}{Classification result.}

\item{cutoff}{The cutoff of confifence score C.}
}
\value{
A heatmap for the confusion matrix of the classification result with the false postive control.
}
\description{
Heatmap for the confusion matrix of the classification with the false postive control.
}
