% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{DsPlot}
\alias{DsPlot}
\title{Point plot for entropy reduction.}
\usage{
DsPlot(ent_res, cutoff = 0, gene_num = 10)
}
\arguments{
\item{ent_res}{The computing result of expression entropy implemented with Entropy function.}

\item{cutoff}{The cutoff of entropy reduction.}

\item{gene_num}{Top genes with maximal entropy reduction.}
}
\value{
A figure.
}
\description{
Point plot for entropy reduction.
}
