% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Entropy}
\alias{Entropy}
\title{Compute expression entropy.}
\usage{
Entropy(expr, window=120, low = 2000)
}
\arguments{
\item{expr}{The expression dataframe. Rows should be cells and columns should be genes.}

\item{window}{The window size for expression value discretization.}

\item{low}{The lower limit for normalizing expression entropy}
}
\value{
A dataframe..
}
\description{
Compute expression entropy.
}
