% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{ExportModel}
\alias{ExportModel}
\title{Export model as matrix.}
\usage{
ExportModel(Bet)
}
\arguments{
\item{Bet}{A prediction function generated by SciBet.}
}
\value{
A matrix.
}
\description{
Export model as matrix.
}
\details{
If you don't plan to export the models for usage on other platforms, simply saving the Bet function in a RData file would also work in R.
}
