% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Learn}
\alias{Learn}
\title{Train SciBet model and generate a "Bet" function for cell type prediction.}
\usage{
Learn(expr, geneset, k=1000, a=5)
}
\arguments{
\item{expr}{The reference expression dataframeThe expression dataframe, with rows being cells, and columns being genes. The last column should be "label".}

\item{geneset}{A user-defined set of genes for prediction.}

\item{k}{See [SelectGene()] for details.}

\item{a}{For each cell type, select a genes by perform E-test in the one-vs-rest manner. If users set a>0, the param k will be neglected.}
}
\value{
A function that takes two inputs: `expr` and `result`. See [Bet()] for details.
}
\description{
Train SciBet model and generate a "Bet" function for cell type prediction.
}
