% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{LoadModel}
\alias{LoadModel}
\title{Generate Bet function from a model matrix.}
\usage{
LoadModel(x, genes, labels)
}
\arguments{
\item{x}{A SciBet model in the format of a matrix.}

\item{genes}{(Optional).}

\item{labels}{(Optional).}
}
\value{
A Bet function.
}
\description{
Generate Bet function from a model matrix.
}
