% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Marker_heatmap}
\alias{Marker_heatmap}
\title{Expression patterns of informative genes across cell types.}
\usage{
Marker_heatmap(expr, gene)
}
\arguments{
\item{expr}{The expression dataframe. Rows should be cells, columns should be genes and last column should be "label".}

\item{gene}{A vector of informative genes.}
}
\value{
A figure.
}
\description{
Expression patterns of informative genes across cell types.
}
