\name{scibet-package}
\alias{scibet-package}
\alias{scibet}
\docType{package}
\title{
  Single-Cell Identifier Based on Entropy Test
}
\description{
  A toolset to apply the SciBet algorithm on single-cell sequencing datasets.
}
\details{

}
\author{
Pauling Liu

Maintainer: Pauling Liu <pauling.liu@pku.edu.cn>
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{
  \dontrun{
    SciBet(train, test)
  }
}
