% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{SelectGene}
\alias{SelectGene}
\title{Select informative genes of the training set in a supervised manner.}
\usage{
SelectGene(expr, k=1000, r = FALSE)
}
\arguments{
\item{expr}{The expression dataframe, with rows being cells, and columns being genes. The last column should be "label".}

\item{k}{Number of genes to select according to the total entropy differneces among cell types.}

\item{r}{Whether to reture a datafram of the gene list. The default is FALSE.}
}
\value{
A list of genes informative genes.
}
\description{
Select informative genes of the training set in a supervised manner.
}
