% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{SEM}
\alias{SEM}
\title{S-E curve.}
\usage{
SEM(ent_res)
}
\arguments{
\item{ent_res}{The computing result of expression entropy implemented with Entropy function.}
}
\value{
A figure.
}
\description{
S-E curve.
}
